+++ packages/contracts/deployer/index.ts
@@ -9,7 +9,12 @@ import * as fs from 'fs';
   if (!proxyAdminOwner)
     throw new Error('Deploy script missing proxy admin owner address');
 
-  const deployment = await deployNocturne(network, proxyAdminOwner);
+  const useMockSubtreeUpdateVerifier =
+    process.env.USE_MOCK_SUBTREE_UPDATE_VERIFIER != undefined;
+
+  const deployment = await deployNocturne(network, proxyAdminOwner, {
+    useMockSubtreeUpdateVerifier,
+  });
   console.log(deployment);
 
   const deploysDir = `${__dirname}/../deploys/`;