+++ packages/sdk/src/sdk/merkleProver/local.ts
@@ -11,16 +11,22 @@ import { fetchInsertions } from "../../indexing";
 const DEFAULT_START_BLOCK = 0;
 const MERKLE_NEXT_BLOCK_TO_INDEX = "MERKLE_NEXT_BLOCK_TO_INDEX";
 
+export interface LocalMerkleProverOpts {
+  startBlock?: number;
+}
+
 export class LocalMerkleProver extends MerkleProver {
   readonly localTree: BinaryPoseidonTree;
   protected contract: Wallet;
   protected provider: ethers.providers.Provider;
   protected db: MerkleDB;
+  protected startBlock: number;
 
   constructor(
     walletContractAddress: Address,
     provider: ethers.providers.Provider,
-    db: MerkleDB
+    db: MerkleDB,
+    opts?: LocalMerkleProverOpts
   ) {
     super();
 
@@ -31,14 +37,16 @@ export class LocalMerkleProver extends MerkleProver {
       this.provider
     );
     this.db = db;
+    this.startBlock = opts?.startBlock ?? DEFAULT_START_BLOCK;
   }
 
   static async fromDb(
     merkleAddress: Address,
     provider: ethers.providers.Provider,
-    db: MerkleDB
+    db: MerkleDB,
+    opts?: LocalMerkleProverOpts
   ): Promise<LocalMerkleProver> {
-    const self = new LocalMerkleProver(merkleAddress, provider, db);
+    const self = new LocalMerkleProver(merkleAddress, provider, db, opts);
 
     for await (const leaf of db.iterLeaves()) {
       self.localTree.insert(leaf);
@@ -63,7 +71,7 @@ export class LocalMerkleProver extends MerkleProver {
     // TODO: load default from network-specific config
     const nextBlockToIndex =
       (await this.db.kv.getNumber(MERKLE_NEXT_BLOCK_TO_INDEX)) ??
-      DEFAULT_START_BLOCK;
+      this.startBlock;
     const latestBlock = await this.provider.getBlockNumber();
 
     const newLeaves = await this.fetchNewLeaves(nextBlockToIndex, latestBlock);