+++ packages/subtree-updater/src/index.ts
@@ -10,7 +10,7 @@ import {
 } from "@nocturne-xyz/sdk";
 import { RootDatabase, Database } from "lmdb";
 import { Wallet } from "@nocturne-xyz/contracts";
-import { SubtreeUpdateSubmitter, SyncSubtreeSubmitter } from "./submitter";
+import { SubtreeUpdateSubmitter } from "./submitter";
 import * as JSON from "bigint-json-serialization";
 
 export { SubtreeUpdateServer } from "./server";
@@ -52,12 +52,14 @@ export class SubtreeUpdater {
   private insertions: (Note | bigint)[];
   private batches: SubtreeUpdateBatch[];
   private tree: BinaryPoseidonTree;
+  private indexingStartBlock: number;
 
   constructor(
     walletContract: Wallet,
     rootDB: RootDatabase,
     prover: SubtreeUpdateProver,
-    submitter: SubtreeUpdateSubmitter = new SyncSubtreeSubmitter(walletContract)
+    submitter: SubtreeUpdateSubmitter,
+    indexingStartBlock: number = 0
   ) {
     this.walletContract = walletContract;
     this.db = rootDB.openDB<string, string>({ name: "insertions" });
@@ -68,6 +70,9 @@ export class SubtreeUpdater {
     this.tree = new BinaryPoseidonTree();
 
     this.submitter = submitter;
+
+    this.indexingStartBlock = indexingStartBlock;
+    console.log("Indexing start block:", this.indexingStartBlock);
   }
 
   public async init(): Promise<void> {
@@ -89,6 +94,10 @@ export class SubtreeUpdater {
     if (nextBlockToIndex > currentBlockNumber) {
       return false;
     }
+
+    console.log(
+      `Indexing. From: ${nextBlockToIndex}. To: ${currentBlockNumber}.`
+    );
 
     const [newInsertions, newCommits] = await Promise.all([
       fetchInsertions(
@@ -172,15 +181,17 @@ export class SubtreeUpdater {
     return madeBatch;
   }
 
+  // TODO: replace plain DB with new put/get numbers in KvDB
   private async getNextBlockToIndex(): Promise<number> {
-    const nextBlockToIndexStr =
-      (await this.db.get(NEXT_BLOCK_TO_INDEX_KEY)) ?? "0";
-    return parseInt(nextBlockToIndexStr);
+    const nextBlockToIndexStr = await this.db.get(NEXT_BLOCK_TO_INDEX_KEY);
+    return nextBlockToIndexStr
+      ? parseInt(nextBlockToIndexStr)
+      : this.indexingStartBlock;
   }
 
   private async getNextInsertionIndex(): Promise<number> {
-    const indexStr = (await this.db.get(NEXT_INSERTION_INDEX_KEY)) ?? "0";
-    return parseInt(indexStr);
+    const indexStr = await this.db.get(NEXT_INSERTION_INDEX_KEY);
+    return indexStr ? parseInt(indexStr) : this.indexingStartBlock;
   }
 
   private async getLastCommittedIndex(): Promise<number | undefined> {