+++ packages/subtree-updater/src/server.ts
@@ -4,8 +4,14 @@ import { SubtreeUpdater } from ".";
 import { Wallet__factory } from "@nocturne-xyz/contracts";
 import { clearTimeout } from "timers";
 import { SubtreeUpdateProver } from "@nocturne-xyz/sdk";
+import { SyncSubtreeSubmitter } from "./submitter";
 
 const TWELVE_SECONDS = 12 * 1000;
+
+export interface SubtreeUpdaterServerOpts {
+  indexingStartBlock?: number;
+  interval?: number;
+}
 
 export class SubtreeUpdateServer {
   private updater: SubtreeUpdater;
@@ -19,7 +25,7 @@ export class SubtreeUpdateServer {
     walletContractAddress: string,
     dbPath: string,
     signer: ethers.Signer,
-    interval: number = TWELVE_SECONDS
+    opts?: SubtreeUpdaterServerOpts
   ) {
     const walletContract = Wallet__factory.connect(
       walletContractAddress,
@@ -27,8 +33,16 @@ export class SubtreeUpdateServer {
     );
     const rootDB = open({ path: dbPath });
 
-    this.interval = interval;
-    this.updater = new SubtreeUpdater(walletContract, rootDB, prover);
+    const submitter = new SyncSubtreeSubmitter(walletContract);
+
+    this.interval = opts?.interval ?? TWELVE_SECONDS;
+    this.updater = new SubtreeUpdater(
+      walletContract,
+      rootDB,
+      prover,
+      submitter,
+      opts?.indexingStartBlock
+    );
     this.stopped = true;
   }