+++ packages/contracts/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+- Deploy script takes mock subtree update verifier param as env var
+- Deploy script logs network info and start block
 - Check if bundler payout > 0 tokens and only call transfer if so
 - fix loop index bug in `extractJoinSplitProofsAndPis`
 - fix contract calculating opDigest differently from frontend