+++ packages/subtree-updater/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+- Add `interval` and `indexingStartBlock` as options to server
+- Remove circuit-artifacts download for mock updater
 - change `build_docker.sh` to avoid `docker buildx` when building mock subtree updater
 - add `yarn build:mock:docker` script and corresponding functionality in `build_docker.sh`
 - create separate subtree updater dockerfile that doesn't use rapidsnark